backend/pyproject.toml
[project]
name = "ai-divination-backend"
version = "0.1.0"
description = "FastAPI backend for AI Divination"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "passlib[bcrypt]>=1.7",
  "redis>=5.0",
  "httpx>=0.27",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0",
  "ruff>=0.3",
  "mypy>=1.8",
]
